fix(flags): Explicitly convert array to parameter string
The behavior of passing an arrayt to `util.format` has changed in Node.js
12. Arrays now include square brackets `[]` around them in the string
representation, which was breaking the parameters passed to the
`pbf2json` go executable.

This change uses `Array.prototype.join` to ensure the array is converted
to a string exactly how we want.

Connects https://github.com/pelias/pelias/issues/800